requirements.txt
beautifulsoup4
genanki
python-docx
requests
